refactor(parser): split lib.rs into smaller logical modules
Refactored the large crates/snail-parser/src/lib.rs file (2394 lines)
into smaller, focused modules:

- util.rs: Span utilities and error handling helpers
- string.rs: String and f-string parsing logic
- literal.rs: Literal and collection parsing (lists, dicts, tuples, sets)
- expr.rs: Expression parsing (binary ops, unary ops, comparisons)
- stmt.rs: Statement parsing (if, while, for, def, class, try, with, etc.)
- awk.rs: AWK-specific parsing (rules, patterns, actions)
- lib.rs: Core parser struct and public API (now ~85 lines)

All modules maintain proper encapsulation and cross-module dependencies.
All tests pass, no warnings from clippy, code is formatted.
